// apps/derrops-portal/src/client/derrops-cloud/models/relay-instance.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Derrops API
 * Internal API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface RelayInstance {
    'id': string;
    'tenant_id': string;
    'name': string;
    'url': string;
    'aegis_id': string | null;
    'status': object;
    'last_seen_at': string | null;
    'created_at': string;
    'updated_at': string;
}